specialized_models_server/requirements.txt
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
anyio==4.0.0
APScheduler==3.10.4
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
bce-python-sdk==0.8.90
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
cattrs==23.2.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.25.0
coloredlogs==15.0.1
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.7
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.2
datasets==2.13.0
decorator==5.1.1
dill==0.3.6
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.13.1
fire==0.5.0
Flask==2.3.3
Flask-APScheduler==1.13.1
flask-babel==3.1.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.12.1
future==0.18.3
gast==0.5.4
GPUtil==1.4.0
h11==0.14.0
h5py==3.10.0
hepunits==2.3.2
hf_transfer==0.1.4
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.4
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
incant==23.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==15.0.7
lmdb==1.4.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
modelscope==1.9.5
mpi4py # @ file:///croot/mpi4py_1671223370575/work
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnx==1.15.0
onnxruntime-gpu==1.12.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
oss2==2.18.3
packaging==23.1
paddle-bfloat==0.1.7
paddleocr==2.7.0.2
paddlepaddle-gpu==2.5.1.post117
pandas==2.1.0
particle==0.23.1
pdf2docx==0.5.6
Pillow==9.4.0
platformdirs==4.1.0
plotly==5.18.0
premailer==3.10.0
protobuf==4.24.2
psutil==5.9.5
pyarrow==14.0.1
pyclipper==1.3.0.post4
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.18.0
PyMuPDF==1.20.2
pynuml==23.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.2.0
rarfile==4.0
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.11.2
seaborn==0.13.0
sentencepiece==0.1.99
shapely==2.0.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sympy==1.12
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.8.30
tokenizers==0.15.0
tomli==2.0.1
torch==2.0.1+cu117
torch_geometric==2.4.0
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.66.1
transformers==4.35.2
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uapi==0.1
urllib3==2.0.4
visualdl==2.5.3
Werkzeug==2.3.7
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.16.2
